# public/assets/api.yaml
openapi: 3.0.0
paths:
...
  /sysDicts:
    get:
      tags:
        - SysDict
      summary: 'Find list SysDict'
      description: 'Returns list of SysDict'
      operationId: getSysDict
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysDict' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysDict' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysDict' } }
                type: object
        '404':
          description: 'SysDict not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysDict
      summary: 'Add a new SysDict to the store'
      operationId: addSysDict
      requestBody:
        $ref: '#/components/requestBodies/SysDict'
      responses:
        '201':
          description: 'Created SysDict'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysDict'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysDict'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysDicts/{id}':
    get:
      tags:
        - SysDict
      summary: 'Find SysDict by ID'
      description: 'Returns a single SysDict'
      operationId: getSysDictById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysDict to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysDict'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysDict'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysDict not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysDict
      summary: 'Update an existing SysDict'
      operationId: updateSysDict
      parameters:
        -
          name: id
          in: path
          description: 'SysDict id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysDict'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysDict not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysDict
      summary: 'Deletes a SysDict'
      operationId: deleteSysDict
      parameters:
        -
          name: id
          in: path
          description: 'SysDict id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /sysMenus:
    get:
      tags:
        - SysMenu
      summary: 'Find list SysMenu'
      description: 'Returns list of SysMenu'
      operationId: getSysMenu
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysMenu' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysMenu' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysMenu' } }
                type: object
        '404':
          description: 'SysMenu not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysMenu
      summary: 'Add a new SysMenu to the store'
      operationId: addSysMenu
      requestBody:
        $ref: '#/components/requestBodies/SysMenu'
      responses:
        '201':
          description: 'Created SysMenu'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysMenu'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysMenu'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysMenus/{id}':
    get:
      tags:
        - SysMenu
      summary: 'Find SysMenu by ID'
      description: 'Returns a single SysMenu'
      operationId: getSysMenuById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysMenu to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysMenu'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysMenu'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysMenu not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysMenu
      summary: 'Update an existing SysMenu'
      operationId: updateSysMenu
      parameters:
        -
          name: id
          in: path
          description: 'SysMenu id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysMenu'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysMenu not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysMenu
      summary: 'Deletes a SysMenu'
      operationId: deleteSysMenu
      parameters:
        -
          name: id
          in: path
          description: 'SysMenu id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    SysDict:
      title: SysDict
      description: SysDict
      properties:
        id:
          nullable: false
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
        name:
          nullable: true
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 50
        code:
          nullable: true
          title: code
          description: code
          type: string
          format: '-'
          maxLength: 50
        status:
          nullable: true
          title: status
          description: status
          type: integer
          format: '-'
          maxLength: 1
        remark:
          nullable: true
          title: remark
          description: remark
          type: string
          format: '-'
          maxLength: 255
        gmt_create:
          nullable: true
          title: gmt_create
          description: gmt_create
          type: string
          format: date
        gmt_modified:
          nullable: true
          title: gmt_modified
          description: gmt_modified
          type: string
          format: date
      type: object
...
    SysMenu:
      title: SysMenu
      description: SysMenu
      properties:
        id:
          nullable: false
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
        name:
          nullable: true
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 64
        parent_id:
          nullable: true
          title: parent_id
          description: parent_id
          type: integer
          format: '-'
          maxLength: 20
        path:
          nullable: true
          title: path
          description: path
          type: string
          format: '-'
          maxLength: 128
        component:
          nullable: true
          title: component
          description: component
          type: string
          format: '-'
          maxLength: 128
        icon:
          nullable: true
          title: icon
          description: icon
          type: string
          format: '-'
          maxLength: 64
        sort:
          nullable: true
          title: sort
          description: sort
          type: integer
          format: '-'
          maxLength: 11
        visible:
          nullable: true
          title: visible
          description: visible
          type: integer
          format: '-'
          maxLength: 1
        redirect:
          nullable: true
          title: redirect
          description: redirect
          type: string
          format: '-'
          maxLength: 128
        gmt_create:
          nullable: true
          title: gmt_create
          description: gmt_create
          type: string
          format: date
        gmt_modified:
          nullable: true
          title: gmt_modified
          description: gmt_modified
          type: string
          format: date
      type: object
...
  requestBodies:
...
    SysDict:
      description: 'SysDict object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysDict'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysDict'
        application/json:
          schema:
            $ref: '#/components/schemas/SysDict'
...
    SysMenu:
      description: 'SysMenu object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysMenu'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysMenu'
        application/json:
          schema:
            $ref: '#/components/schemas/SysMenu'
...